third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.cookies-allowed.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins cookie access when third-party cookie access allowed
  assert_equals: expected "ReadOnLoad:FirstPartyNone-ThirdPartyNone-,ReadOnFetch:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-" but got "ReadOnLoad:FirstPartyLax-FirstPartyNone-,ReadOnFetch:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-"
Harness: the test ran to completion.
